+++ app/Enum/MediaTypes.ts
@@ -8,4 +8,5 @@
 export class MediaTypes {
   public static APPLICATION_JSON = 'application/json';
   public static FORM_URLENCODED = 'application/x-www-form-urlencoded';
+  public static TEXT_PLAIN = 'text/plain';
 }