+++ test/mocks/MockInterceptor.ts
@@ -1,0 +1,4 @@
+import {HttprInterceptor} from '../../app/Core/HttprInterceptor';
+
+export class MockInterceptor extends HttprInterceptor {
+}