+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Added travis integration.
 - Added LICENSE file.
+- Added `text/plain` media type.
+- Added tests.
+- Fixed missing body field for non JSON types.
 
 ## 1.0.0-rc2